src/player: Fix compilation error with gstreamer
* src/player/NewVideoPlayer.cpp
* src/player/NewVideoPlayer.h
* tests-ncl/test-image-anchor-begin-big.ncl: New test

+++ src/player/NewVideoPlayer.h
@@ -67,7 +67,7 @@ public:
   int64_t getVPts ();
   void timeShift (const string &direction);
 
-  double getMediaTime ();
+  guint32 getMediaTime ();
   void setMediaTime (double pos);
   void setStopTime (double pos);
   double getStopTime ();
